requirements.txt
boto3==1.10.2
botocore==1.13.2
Django==2.2.6
docutils==0.15.2
image==1.5.27
jmespath==0.9.4
Pillow==6.2.1
python-dateutil==2.8.0
pytz==2019.3
s3transfer==0.2.1
six==1.12.0
sqlparse==0.3.0
urllib3==1.25.6
